Add Open from Feed menu which opens an empty dialog.
The Open from Feed menu opens a modal dialog. Opening the dialog
is just a case of using:

NSApplication.SharedApplication.RunModalForWindow (controller.Window);

Passing the window for the dialog.

Getting the Close button or the window close button to close the
modal was difficult. Having an action linked to a Close button on
the window would close the main window if the StopModal method was
used:

NSApplication.SharedApplication.StopModal ();

Also the window's red close button would hide the modal dialog but
not return control back to the main window so it was not possible
to quit the application.

The workaround seems to be to have the modal window use its
controller as its delegate. Then the standard windowWillClose:
action can be caught by the modal window's controller and then
StopModal can be called. Having a modal window use it controller
as its delegate you have to Ctrl+Click the window in Xcode and drag
a line over to the File's Owner. A perhaps more obvious way of doing
this is to open the Connections Inspector, find delegate in the
Outlets, ctrl+click the circle and drag that to the File's Owner.